feat: working cart w/ context and sub-comps
