+++ src/features/shopping-cart/components/add-item/index.ts
@@ -1,0 +1,1 @@
+export { AddItem } from './add-item';
