+++ src/features/shopping-cart/components/cart/cart.module.scss
