+++ src/features/shopping-cart/components/cart/index.ts
@@ -1,0 +1,1 @@
+export { Cart } from './cart';
