+++ src/features/shopping-cart/shopping-cart.tsx
@@ -1,211 +1,243 @@
 import React from 'react';
-import { useRouter } from 'next/router';
-import { useCreatePaymentSession } from '@modules/queries/payments';
-import { useCreateCart, useSyncCartItems } from '@modules/queries/carts';
-import { isAuthenticated } from '@modules/queries/auth';
-import { useCheckout } from '@components/hooks/use-checkout';
-import { ShoppingCartProps } from '@components/hooks/use-shopping-cart';
-import { Cart, Button, Drawer, Incrementor, Alert } from '@components/app';
-import { AlertError, Trash } from '@components/icons';
-import { ServerErrors } from '@components/server-error';
-import { ShoppingBag } from '@components/illustrations';
-import { Error } from '@modules/api-response/typings';
-import styles from './shopping-cart.module.scss';
+import { GetProduct } from '@typings/stripe';
+import { Cart } from './components/cart';
+import { AddItem } from './components/add-item';
+import { ShoppingCartContext } from './shopping-cart-context';
 
-export const ShoppingCart = ({
-  items,
-  quantity,
-  total,
-  incrementItem,
-  decrementItem,
-  updateCart,
-  removeCartItem,
-}: Partial<ShoppingCartProps>) => {
-  const router = useRouter();
-  const [open, setOpen] = React.useState(false);
-  const [checkingOut, setCheckingOut] = React.useState(false);
-  const [serverErrors, setServerErrors] = React.useState<Error[]>([]);
+type CartState = {
+  items: GetProduct[];
+  total: number;
+  status: string;
+};
 
-  const [createPaymentSession] = useCreatePaymentSession();
-  const [createCart] = useCreateCart();
-  const [syncCartItems] = useSyncCartItems({
-    onSuccess: (result) => {
-      if (updateCart) updateCart(result.data);
-    },
-  });
+type ShoppingCart = {
+  children: React.ReactNode;
+};
 
-  const handleCheckout = async () => {
-    createCart(
-      {},
-      {
-        onSuccess: (createdCart) => {
-          syncCartItems({
-            cartId: createdCart.data.attributes.id || '',
-            cartItems: items || [],
-          });
-        },
-      }
-    );
+const ShoppingCart = ({ children }: ShoppingCart) => {
+  const initialState = { items: [], total: 0, status: 'new' };
 
-    const response = await createPaymentSession({
-      orderItems: items || [],
-    });
+  if (typeof window === 'undefined') {
+    return null;
+  }
 
-    const checkoutError = await useCheckout({
-      sessionId: response.data.attributes.id,
-    });
+  const storage = window.localStorage;
+  const [state, setState] = React.useState<CartState>(initialState);
 
-    if (checkoutError) {
-      return setServerErrors([{ ...checkoutError }]);
-    }
+  const getCartTotal = (items: GetProduct[] = []) => {
+    return items.reduce((acc: number, item: GetProduct) => {
+      let result = acc;
 
-    return undefined;
+      result +=
+        ((item.relationships?.price.unit_amount || 0) *
+          item.attributes.quantity) /
+        100;
+
+      return Number(result.toFixed(2));
+    }, 0);
   };
 
-  isAuthenticated({
-    enabled: checkingOut,
-    onSuccess: async () => {
-      await handleCheckout();
-    },
-    onError: (error) => {
-      return error?.response?.data?.error.map((e: Error) => {
-        if (e.code === 'UNAUTHENTICATED') {
-          return router.push(
-            '/app/login?return_to=checkout',
-            '/app/login?return_to=checkout'
-          );
-        }
+  const addCartItem = (item: GetProduct) => {
+    const storageCart = storage.getItem('cart') || '{}';
+    const currentCart: CartState = JSON.parse(storageCart);
+    let newCartItems: GetProduct[] = [];
 
-        return undefined;
+    /* See if the items already exists */
+    const existingItem = currentCart.items.find(
+      (i) => i.attributes.id === item.attributes.id
+    );
+
+    if (existingItem) {
+      existingItem.attributes.quantity += 1;
+    } else {
+      newCartItems = currentCart.items.concat({
+        ...item,
+        attributes: { ...item.attributes, quantity: 1 },
       });
-    },
-  });
+    }
+
+    const newCartTotal = getCartTotal(
+      existingItem ? currentCart.items : newCartItems
+    );
+
+    storage.setItem(
+      'cart',
+      JSON.stringify({
+        ...currentCart,
+        items: existingItem ? currentCart.items : newCartItems,
+        total: newCartTotal,
+        status: 'active',
+      })
+    );
+
+    setState({
+      ...state,
+      items: existingItem ? currentCart.items : newCartItems,
+      total: newCartTotal,
+    });
+  };
+
+  const incrementItem = (item: GetProduct) => {
+    const storageCart = storage.getItem('cart') || '{}';
+    const currentCart: CartState = JSON.parse(storageCart);
+
+    /* Find the existing item to increment */
+    const existingItem = currentCart.items.find(
+      (i) => i.attributes.id === item.attributes.id
+    );
+
+    if (existingItem) {
+      existingItem.attributes.quantity += 1;
+
+      const newCartTotal = getCartTotal(currentCart.items);
+
+      storage.setItem(
+        'cart',
+        JSON.stringify({
+          ...currentCart,
+          items: currentCart.items,
+          total: newCartTotal,
+          status: 'active',
+        })
+      );
+
+      setState({
+        ...state,
+        items: currentCart.items,
+        total: newCartTotal,
+      });
+    }
+  };
+
+  const decrementItem = (item: GetProduct) => {
+    const storageCart = storage.getItem('cart') || '{}';
+    const currentCart: CartState = JSON.parse(storageCart);
+
+    /* Get the existing item to decrement */
+    const existingItem = currentCart.items.find(
+      (i) => i.attributes.id === item.attributes.id
+    );
+
+    if (existingItem) {
+      existingItem.attributes.quantity -= 1;
+
+      const newCartTotal = getCartTotal(currentCart.items);
+
+      storage.setItem(
+        'cart',
+        JSON.stringify({
+          ...currentCart,
+          items: currentCart.items,
+        })
+      );
+
+      setState({
+        ...state,
+        items: currentCart.items,
+        total: newCartTotal,
+      });
+    }
+  };
+
+  const removeCartItem = (item: GetProduct) => {
+    const storageCart = storage.getItem('cart') || '{}';
+    const currentCart: CartState = JSON.parse(storageCart);
+
+    const newCartItems = currentCart.items.filter(
+      (product: GetProduct) => item.attributes.id !== product.attributes.id
+    );
+
+    const cartStatus = newCartItems.length > 0 ? 'active' : 'new';
+
+    const newCartTotal = getCartTotal(newCartItems);
+
+    storage.setItem(
+      'cart',
+      JSON.stringify({
+        ...currentCart,
+        items: newCartItems,
+        status: cartStatus,
+      })
+    );
+
+    setState({
+      ...state,
+      items: newCartItems,
+      total: newCartTotal,
+      status: cartStatus,
+    });
+  };
+
+  const updateCart = (items: GetProduct[]) => {
+    const newCartTotal = getCartTotal(items);
+
+    storage.setItem(
+      'cart',
+      JSON.stringify({
+        items,
+        total: newCartTotal,
+        status: 'active',
+      })
+    );
+
+    setState({
+      ...state,
+      items,
+      total: newCartTotal,
+    });
+  };
+
+  const clearCart = () => {
+    storage.removeItem('cart');
+  };
+
+  const calculateQuantity = (items: GetProduct[]) => {
+    return items?.reduce((acc, item) => {
+      let result = acc;
+      result += item.attributes?.quantity;
+      return result;
+    }, 0);
+  };
+
+  React.useEffect(() => {
+    const storageCart = storage.getItem('cart') || '{}';
+    const currentCart: CartState = JSON.parse(storageCart);
+    const emptyCart = !Object.keys(currentCart).length;
+
+    if (emptyCart) {
+      storage.setItem('cart', JSON.stringify(initialState));
+    } else {
+      setState({
+        items: currentCart.items,
+        total: getCartTotal(currentCart.items),
+        status: currentCart.status,
+      });
+    }
+  }, []);
 
   return (
-    <>
-      <Cart count={quantity || 0} onClick={() => setOpen(true)} />
-
-      {/* Drawer */}
-      <Drawer
-        isOpen={open}
-        onClose={() => setOpen(false)}
-        className="rounded-tl-md rounded-bl-md"
-        width="24rem"
-      >
-        <div className="flex flex-col h-full relative px-5 py-6 space-y-3">
-          <div className="relative">
-            <div className="flex-1 text-xl text-center font-medium leading-tight">
-              My Cart
-            </div>
-            <Drawer.Close className={styles.close} />
-          </div>
-
-          <ul className="flex-1 divide-y last:divide-y-0 divide-gray-200">
-            {!items?.length && (
-              <div className="flex flex-col justify-center items-center space-y-6 h-full">
-                <ShoppingBag size={225} className="mx-auto" />
-                <div className="text-center font-medium mt-8">
-                  Your cart is currently empty.
-                </div>
-                <Button variant="primary" onClick={() => setOpen(false)}>
-                  Continue shopping
-                </Button>
-              </div>
-            )}
-
-            {items?.map((item) => {
-              return (
-                <li key={item.attributes.id} className="py-3">
-                  <div className="flex justify-between space-x-3">
-                    <div>
-                      <img
-                        src={`/images/products/${item.attributes.id}.jpg`}
-                        alt={item.attributes.name}
-                        className="h-16 w-16 rounded-md"
-                      />
-                    </div>
-                    <div className="flex-1">
-                      <div className="flex flex-col space-y-2">
-                        <div className="flex justify-between space-x-8">
-                          <div>{item.attributes.name}</div>
-                          <div className="font-medium">
-                            {(
-                              (item.relationships?.price.unit_amount || 0) / 100
-                            ).toLocaleString('en-US', {
-                              style: 'currency',
-                              currency: 'USD',
-                            })}
-                          </div>
-                        </div>
-
-                        <div className="flex items-center justify-between">
-                          <div>
-                            <Incrementor
-                              value={item.attributes.quantity}
-                              min={1}
-                              onIncrement={() => {
-                                if (incrementItem) incrementItem(item);
-                              }}
-                              onDecrement={() => {
-                                if (decrementItem) decrementItem(item);
-                              }}
-                            />
-                          </div>
-
-                          <div className={styles.trash}>
-                            <Trash
-                              size={18}
-                              onClick={() => {
-                                if (removeCartItem) removeCartItem(item);
-                              }}
-                            />
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </li>
-              );
-            })}
-          </ul>
-
-          {(quantity || 0) > 0 && (
-            <>
-              <div className="flex items-center justify-between font-medium pt-2">
-                <div>Item Total</div>
-
-                <div>
-                  {total?.toLocaleString('en-US', {
-                    style: 'currency',
-                    currency: 'USD',
-                  })}
-                </div>
-              </div>
-
-              {serverErrors.length > 0 && (
-                <Alert variant="error" className="mb-4">
-                  <div className="mr-3">
-                    <AlertError size={18} />
-                  </div>
-                  <Alert.Content>
-                    <ServerErrors errors={serverErrors} />
-                  </Alert.Content>
-                </Alert>
-              )}
-
-              <div className="flex justify-center">
-                <Button
-                  variant="primary"
-                  className="w-11/12"
-                  onClick={() => setCheckingOut(true)}
-                >
-                  Checkout
-                </Button>
-              </div>
-            </>
-          )}
-        </div>
-      </Drawer>
-    </>
+    <ShoppingCartContext.Provider
+      value={{
+        items: state.items,
+        quantity: calculateQuantity(state.items),
+        total: state.total,
+        status: state.status,
+        addCartItem,
+        updateCart,
+        incrementItem,
+        decrementItem,
+        removeCartItem,
+        clearCart,
+        calculateQuantity,
+      }}
+    >
+      {children}
+    </ShoppingCartContext.Provider>
   );
 };
+
+// Sub-components
+ShoppingCart.Cart = Cart;
+ShoppingCart.AddItem = AddItem;
+
+export { ShoppingCart };
