+++ src/views/app/products/products.tsx
@@ -1,52 +1,26 @@
 import React from 'react';
 import { useGetProducts } from '@modules/queries/products';
-import { useShoppingCart } from '@components/hooks/use-shopping-cart';
-import { useActor } from '@components/hooks/use-actor';
-import { useGetActorCart } from '@modules/queries/actor';
-import { Button } from '@components/app';
+// import { useActor } from '@components/hooks/use-actor';
+// import { useGetActorCart } from '@modules/queries/actor';
+// import { Button } from '@components/app';
 import { ShoppingCart } from '@features/shopping-cart';
 // import styles from './products.module.scss';
 
 const Products = () => {
-  const {
-    items,
-    quantity,
-    total,
-    addCartItem,
-    removeCartItem,
-    updateCart,
-    incrementItem,
-    decrementItem,
-    calculateQuantity,
-  } = useShoppingCart();
-  const [actorId] = useActor();
+  // const [actorId] = useActor();
   const { data: response, isLoading } = useGetProducts();
-  const { data: cartItems } = useGetActorCart(actorId, {
-    enabled: !!actorId,
-  });
+  // const { data: cartItems } = useGetActorCart(actorId, {
+  //   enabled: !!actorId,
+  // });
 
   const products = response?.data;
-  const actorCartItems = cartItems?.data;
-  const shoppingCartItems = actorId ? actorCartItems : items;
-  const shoppingCartQuantity = actorId
-    ? calculateQuantity(actorCartItems)
-    : quantity;
+  // const actorCartItems = cartItems?.data;
 
   return (
-    <div>
+    <ShoppingCart>
       <div className="my-4 mx-8">
         <div className="flex justify-end mb-4">
-          <ShoppingCart
-            items={shoppingCartItems}
-            quantity={shoppingCartQuantity}
-            total={total}
-            incrementItem={(item) => incrementItem(item)}
-            decrementItem={(item) => {
-              decrementItem(item);
-            }}
-            updateCart={updateCart}
-            removeCartItem={removeCartItem}
-          />
+          <ShoppingCart.Cart />
         </div>
         <div className="grid grid-cols-4 gap-4">
           {!isLoading &&
@@ -68,13 +42,14 @@ const Products = () => {
                       currency: 'USD',
                     })}
                   </div>
-                  <Button onClick={() => addCartItem(result)}>Buy Now</Button>
+
+                  <ShoppingCart.AddItem item={result} />
                 </div>
               );
             })}
         </div>
       </div>
-    </div>
+    </ShoppingCart>
   );
 };
